Server.js
I decided to add a server file but Ive found that I could combine this file with the connection file to be an index file that accomplishes both functionalities. To save the trouble of getting confused and staying in line with the MVC file structure im keeping the server file.