Models
Ive added the models for the User and Thoughts based on the structure provided in the course material. However the getter portion of the models im confused about. I need to find a way to include a date function so I can establuish the current timestap for the createdAt portion of the model format in the THoughts Model.