Moment.js
In order to use the getter method to show a timestamp I used moment.js as we did in a previous chalenge. I saw in some other example that you can create a helper.js file to format the date but the mongodb documentation showed a much easier way of dsiplay time using moment.