userRoutes
Following the student mini project, I created the routes for the user. I based the structure on the requirements from the course material and followed how the student mini project coded its routes to code mine. This includes get, post, put, and delete for user and friends based on themselves and ids.