getThoughts
Im now adding the controllers for thr thoughtsController file. Im using the code I used for the userController file to complete this so most of the code is pretty similar with few alterations such as dbUserData will be dbThoughtData and the models will be thought based. Also I will need to add the reactions for the thoughts.