createThought
it took some time to figure out this controller as when we create the thought we have to add it to a designated user id while also maintianing the current body that was added. I used the addFriend controller from the user controllers to model it after as it uses the same methods to add to the id while taking in the parameters of the friends but in this case its the body of the thought.