deleteThought, createReaction, deleteReaction
I finished up the rest of the thought controllers by being able to delete thoughts. THis was a little differetn as you have to update the user associated with the thought so i used the updateUser to help structure the controller. The create and elete reactions were very similar to the friends based in the user controllers.